Move file location and remove unnecessary methods
