+++ src/Main/InterfaceLayer/Document.ts
@@ -1,44 +1,51 @@
-import { DocsInlineObjectsModel } from "../../Models/DocsDocumentModel"
-import { MonadDefinitions } from "../../Utility/Interfaces"
-import { List } from "../../Utility/List"
-import { ListApplicativeWrapperFn, List_2D } from "../../Utility/List_2D"
+// import { DocsInlineObjectsModel } from "../Models/DocsDocumentModel"
+// import { MonadDefinitions } from "../../Utility/Interfaces"
+// import { List } from "../../Utility/List"
+// import { ListApplicativeWrapperFn, List_2D } from "../../Utility/List_2D"
 
-interface Applicative<Value> extends MonadDefinitions.Applicative<Value>{
-}
+// interface Applicative<Value> extends MonadDefinitions.Applicative<Value>{
+// }
 
-export class Document<BaseElement>{
-    $value: List_2D<BaseElement>
-    inlineObj?: DocsInlineObjectsModel
+// interface IDocument<BaseElement>{
+//     $value: List_2D<BaseElement>
+//     inlineObj?: DocsInlineObjectsModel
+//     compactMapWithInlineObj<B>(fn:(inlineObj?:DocsInlineObjectsModel) => (value:BaseElement) => B, baseElementPredicate?:(x:B) => boolean): IDocument<B>
+//     sequence<A, B extends Applicative<Document<A>>>(this:Document<Applicative<A>>, of: ListApplicativeWrapperFn<A>, returnType?:(arg:Document<A>) => B): B
+// }
 
-    constructor(paragraphs: List_2D<BaseElement>, inlineObj?:DocsInlineObjectsModel ){
-        this.$value = paragraphs
-        this.inlineObj = inlineObj
-    }
+// export class Document<BaseElement>{
+//     $value: List_2D<BaseElement>
+//     inlineObj?: DocsInlineObjectsModel
 
-    static of<A>(paragraphs:List_2D<A>, inlineObj?:DocsInlineObjectsModel):Document<A>{
-        return new Document(paragraphs, inlineObj)
-    }
+//     constructor(paragraphs: List_2D<BaseElement>, inlineObj?:DocsInlineObjectsModel ){
+//         this.$value = paragraphs
+//         this.inlineObj = inlineObj
+//     }
 
-    static from2DArray<A>(paragraphs:Array<Array<A>>, inlineObj?: DocsInlineObjectsModel):Document<A>{
-        return Document.of( List_2D.from2DArr(paragraphs), inlineObj)
-    }
-    static fromList<A>(paragraphs:List<List<A>>, inlineObj?: DocsInlineObjectsModel):Document<A>{
-        return Document.of( List_2D.of(paragraphs), inlineObj)
-    }
+//     static of<A>(paragraphs:List_2D<A>, inlineObj?:DocsInlineObjectsModel):Document<A>{
+//         return new Document(paragraphs, inlineObj)
+//     }
 
-    compactMap<B>(fn:(value:BaseElement) => B, baseElementPredicate?:(x:B) => boolean): Document<B>{
-        return Document.of( this.$value.compactMap(fn, baseElementPredicate), this.inlineObj )
-    }
+//     static from2DArray<A>(paragraphs:Array<Array<A>>, inlineObj?: DocsInlineObjectsModel):Document<A>{
+//         return Document.of( List_2D.from2DArr(paragraphs), inlineObj)
+//     }
+//     static fromList<A>(paragraphs:List<List<A>>, inlineObj?: DocsInlineObjectsModel):Document<A>{
+//         return Document.of( List_2D.of(paragraphs), inlineObj)
+//     }
 
-    compactMapWithInlineObj<B>(fn:(inlineObj?:DocsInlineObjectsModel) => (value:BaseElement) => B, baseElementPredicate?:(x:B) => boolean): Document<B>{
-        return this.compactMap(fn(this.inlineObj), baseElementPredicate)
-    }
+//     compactMap<B>(fn:(value:BaseElement) => B, baseElementPredicate?:(x:B) => boolean): Document<B>{
+//         return Document.of( this.$value.compactMap(fn, baseElementPredicate), this.inlineObj )
+//     }
 
-    sequence<A, B extends Applicative<Document<A>>>(this:Document<Applicative<A>>, of: ListApplicativeWrapperFn<A>, returnType?:(arg:Document<A>) => B): B{
-        return this.$value.sequence(of).map(list_2D => Document.of(list_2D, this.inlineObj)) as B
-    }
+//     compactMapWithInlineObj<B>(fn:(inlineObj?:DocsInlineObjectsModel) => (value:BaseElement) => B, baseElementPredicate?:(x:B) => boolean): Document<B>{
+//         return this.compactMap(fn(this.inlineObj), baseElementPredicate)
+//     }
+
+//     sequence<A, B extends Applicative<Document<A>>>(this:Document<Applicative<A>>, of: ListApplicativeWrapperFn<A>, returnType?:(arg:Document<A>) => B): B{
+//         return this.$value.sequence(of).map(list_2D => Document.of(list_2D, this.inlineObj)) as B
+//     }
     
-    reduce<A>(fn:(initialValue:A, currentValue:List<BaseElement>) => A, initialValue:A):A{
-        return this.$value.reduce(fn, initialValue)
-    }
-}
+//     reduce<A>(fn:(initialValue:A, currentValue:List<BaseElement>) => A, initialValue:A):A{
+//         return this.$value.reduce(fn, initialValue)
+//     }
+// }
